+++ packages/devtools_app/lib/src/screens/memory/framework/connected/memory_controller.dart
@@ -6,9 +6,7 @@ import 'dart:async';
 
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/foundation.dart';
-import 'package:vm_service/vm_service.dart';
 
-import '../../../../shared/globals.dart';
 import '../../../../shared/memory/class_name.dart';
 import '../../../../shared/memory/heap_graph_loader.dart';
 import '../../panes/chart/controller/chart_pane_controller.dart';
@@ -19,8 +17,6 @@ import '../../panes/control/controller/control_pane_controller.dart';
 import '../../panes/diff/controller/diff_pane_controller.dart';
 import '../../panes/profile/profile_pane_controller.dart';
 import '../../panes/tracing/tracing_pane_controller.dart';
-import '../../shared/primitives/memory_timeline.dart';
-import 'memory_tracker.dart';
 
 class MemoryFeatureControllers {
   /// Controllers are passed for testability.
@@ -120,104 +116,6 @@ class MemoryController extends DisposableController
   /// DevTools screen changes, so we must store this value in the controller
   /// instead of the widget state.
   int selectedFeatureTabIndex = 0;
-
-  void _handleConnectionStart() {
-    if (controllers.chart.memoryTracker == null) {
-      controllers.chart.memoryTracker =
-          MemoryTracker(controllers.chart.memoryTimeline, controllers.chart);
-      controllers.chart.memoryTracker!.start();
-    }
-
-    // Log Flutter extension events.
-    // Note: We do not need to listen to event history here because we do not
-    // have matching historical data about total memory usage.
-    autoDisposeStreamSubscription(
-      serviceConnection.serviceManager.service!.onExtensionEvent
-          .listen((Event event) {
-        var extensionEventKind = event.extensionKind;
-        String? customEventKind;
-        if (MemoryTimeline.isCustomEvent(event.extensionKind!)) {
-          extensionEventKind = MemoryTimeline.devToolsExtensionEvent;
-          customEventKind =
-              MemoryTimeline.customEventName(event.extensionKind!);
-        }
-        final jsonData = event.extensionData!.data.cast<String, Object>();
-        // TODO(terry): Display events enabled in a settings page for now only these events.
-        switch (extensionEventKind) {
-          case 'Flutter.ImageSizesForFrame':
-            controllers.chart.memoryTimeline.addExtensionEvent(
-              event.timestamp,
-              event.extensionKind,
-              jsonData,
-            );
-            break;
-          case MemoryTimeline.devToolsExtensionEvent:
-            controllers.chart.memoryTimeline.addExtensionEvent(
-              event.timestamp,
-              MemoryTimeline.customDevToolsEvent,
-              jsonData,
-              customEventName: customEventKind,
-            );
-            break;
-        }
-      }),
-    );
-
-    autoDisposeStreamSubscription(
-      controllers.chart.memoryTracker!.onChange.listen((_) {
-        controllers.chart.memoryTrackerController
-            .add(controllers.chart.memoryTracker);
-      }),
-    );
-    autoDisposeStreamSubscription(
-      controllers.chart.memoryTracker!.onChange.listen((_) {
-        controllers.chart.memoryTrackerController
-            .add(controllers.chart.memoryTracker);
-      }),
-    );
-
-    // TODO(terry): Used to detect stream being closed from the
-    // memoryController dispose method.  Needed when a HOT RELOAD
-    // will call dispose however, initState doesn't seem
-    // to happen David is working on scaffolding.
-    controllers.chart.memoryTrackerController.stream.listen(
-      (_) {},
-      onDone: () {
-        // Stop polling and reset memoryTracker.
-        controllers.chart.memoryTracker?.stop();
-        controllers.chart.memoryTracker = null;
-      },
-    );
-
-    controllers.chart.updateAndroidChartVisibility();
-    addAutoDisposeListener(
-      preferences.memory.androidCollectionEnabled,
-      controllers.chart.updateAndroidChartVisibility,
-    );
-  }
-
-  void _handleConnectionStop() {
-    controllers.chart.memoryTracker?.stop();
-    controllers.chart.memoryTrackerController
-        .add(controllers.chart.memoryTracker);
-
-    controllers.reset();
-    controllers.chart.hasStopped = true;
-  }
-
-  void startTimeline() {
-    addAutoDisposeListener(serviceConnection.serviceManager.connectedState, () {
-      if (serviceConnection.serviceManager.connectedState.value.connected) {
-        _handleConnectionStart();
-      } else {
-        _handleConnectionStop();
-      }
-    });
-
-    if (serviceConnection.serviceManager.connectedAppInitialized) {
-      _handleConnectionStart();
-    }
-  }
 
   @override
   void dispose() {
